Extract generic AppModel code.
